Feat. change fields state into atom via jotai
