+++ src/components/Modals/AddNewDocument/AddDocInputList.jsx
@@ -1,18 +1,31 @@
 import { useQuery } from "@tanstack/react-query";
-import { useAtomValue } from "jotai";
-import PropTypes from "prop-types";
+import { useAtomValue, useAtom } from "jotai";
 
 import fetchData from "../../../utils/axios";
 import getTodaysDate from "../../../utils/getTodaysDate";
 
-import { currentDBIdAtom, userAtom } from "../../../atoms/atoms";
+import { currentDBIdAtom, userAtom, fieldsAtom } from "../../../atoms/atoms";
 
 import InputWrapper from "../SharedItems/InputWrapper";
 import Loading from "../../shared/Loading";
 
-function AddDocInputList({ updateFieldValue, setFields }) {
+function AddDocInputList() {
+  const [fields, setFields] = useAtom(fieldsAtom);
   const { userId } = useAtomValue(userAtom);
   const currentDBId = useAtomValue(currentDBIdAtom);
+
+  function adjustTextareaHeight(event) {
+    event.target.style.height = `${event.target.scrollHeight}px`;
+  }
+
+  function updateFieldValue(index, event) {
+    const newFields = [...fields];
+
+    newFields[index].fieldValue = event.target.value;
+
+    setFields(newFields);
+    adjustTextareaHeight(event);
+  }
 
   async function getDatabase() {
     const response = await fetchData(
@@ -75,10 +88,5 @@ function AddDocInputList({ updateFieldValue, setFields }) {
     );
   });
 }
-
-AddDocInputList.propTypes = {
-  updateFieldValue: PropTypes.func.isRequired,
-  setFields: PropTypes.func.isRequired,
-};
 
 export default AddDocInputList;
